+++ src/main/kotlin/engine/shape/Shape2D.kt
@@ -2,5 +2,5 @@ package org.soyuz.engine.shape
 
 sealed interface Shape2D {
     fun localAabb(): Aabb2D
-    fun supportRadius(): Float
+    fun supportRadius(): Double
 }
